+++ src/Accelerators/NNPA/Conversion/ZHighToZLow/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_onnx_mlir_library(OMZHighToZLow
   ProcessStickData.cpp
+  ProcessStickDataHelper.cpp
   ZHighToZLow.cpp
 
   LINK_LIBS PUBLIC
